# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Suthar Enterprises ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/references/add": {
            "post": {
                "tags": [
                    "References APIs"
                ],
                "summary": "Add Reference",
                "description": "Add Reference",
                "operationId": "references",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "full_name",
                                    "phone_number",
                                    "email",
                                    "building_type",
                                    "state",
                                    "city",
                                    "address",
                                    "area_in_sqft",
                                    "frames_count",
                                    "store_id"
                                ],
                                "properties": {
                                    "full_name": {
                                        "type": "text"
                                    },
                                    "phone_number": {
                                        "type": "text"
                                    },
                                    "email": {
                                        "type": "email"
                                    },
                                    "building_type": {
                                        "type": "text"
                                    },
                                    "state": {
                                        "type": "text"
                                    },
                                    "city": {
                                        "type": "text"
                                    },
                                    "address": {
                                        "type": "text"
                                    },
                                    "landmark": {
                                        "type": "text"
                                    },
                                    "area_in_sqft": {
                                        "type": "integer"
                                    },
                                    "frames_count": {
                                        "type": "integer"
                                    },
                                    "store_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reference added successfully.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}